Proper Primary Health Endpoint for Standby Leader
When initializing a standby cluster, the 'standby_leader' REST endpoint
(as Provided by Patroni) is now utilized to determine if the local
instance has initialized as the leader of a standby cluster.  When
initializing a non-standby cluster, the 'master' REST endpoint is still
utilized for this purpose.

This ensures proper leader detection for standby and non-standby
clusters, specifically due to a change in behavior (listed as a
bug fix) in Patroni 2.0.0 in which the 'master' endpoint can no longer
be utilized to detect whether or not the local instance is running
as a standby leader.

Issue: https://github.com/CrunchyData/postgres-operator/issues/1959