Mark PostgreSQL's data volume to detect failed init
When bootstrap fails, Patroni renames the data directory and exits. The
next time the container starts, the "missing" data directory triggers a
successful `initdb` bootstrap, and the cluster comes online empty. When
initializing from an existing data directory, being online like this can
lead to future data loss.

To detect bootstrap failures across restarts, we place a file next to
the PostgreSQL data directory and remove it when bootstrap succeeds. If
we detect this file during startup, we pause the process (logging in a
loop) and do not proceed.

Some human intervention is now required whenever bootstrap fails. This
could be relaxed in the future to allow for some automated retry.

Issue: [ch9711]