crunchy-pgBackRest container
This update compacts the functionality of the existing
crunchy-backrest-restore container along with the
pgo-backrest and pgo-backrest-repo containers from the
Crunchy Postgres Operator project into a main container,
crunchy-pgbackrest, as well as a subsequent container,
crunchy-pgbackrest-repo, that allows for a separate user
as currently used by the existing pgBackRest implementation.

The various pgBackRest operating functions from the existing
containers are now accomplished by the use of operating modes.
These modes are as follows:

pgbackrest mode is used for taking pgBackRest backups, retrieving
info and stanza creation.

pgbackrest-repo mode acts as a pgBackRest remote repository for
the Postgres cluster to use for storing archive files and backups.

pgbackrest-restore mode executes a pgBackRest restore independent
of the Crunchy PostgreSQL Operator.

Finally, as mentioned above, the pgbackrest-repo container allows
for the operation in the 'pgbackrest-repo' mode, but with a
separate, pgbackrest user.

The final container incorporates all necessary scripts, configurations
and code necessary to execute the various modes, while removing any
redundant or otherwise unnecessary dependencies. A new script,
start.sh, determines the correct operating mode based on the value
in the MODE environment variable, and then calls the correct subset
of the overall functionality.