Containers Compaction: pgBouncer pgBadger
In an effort to remove the crunchy-pg-base image the crunchy-pgbouncer and
crunchy-pgbadger iamges are now built based on the crunchy-base image. Since
neither pgBouncer or pgBadger depend on having PostgreSQL installed, using the
crunchy-pg-base image was making the images larger than they needed to be. This
also removes two more dependencies on the crunchy-pg-base image.

The scripts for crunchy-pgbouncer and crunchy-pgbadger have been updated with
the new `/opt/crunchy` (previously `/opt/cpm`) path when adding binary and
config files. The dockerfiles for these two images now include steps to setup
the required repository files and packager settings based on the baseos. These
steps were previously done in the crunchy-pg-base image.

The pgBouncer dockerfile was installing an epel version of pgBouncer instead of
the package provided by Crunchy. When installing pgBouncer the epel repo is now
disabled.