Remove bashrc and bash_profile
These files were used to set env variables when logging into the container.
The variables are now set in the various scripts so the bash files are being
removed as part of this change.

There are still references to these files in these files that will need to be
removed.

build/postgres-appdev/Dockerfile
build/backrest-restore/Dockerfile