Enable Delta Restores When Using pgbackrest_init
If a pgBackRest restore has been requested when bootstrapping a new
PostgreSQL cluster (i.e. the PGHA_BOOTSTRAP_METHOD environment
variable is set to 'pgbackrest_init'), and the PGDATA directory
contains an existing PostgreSQL database, then a pgBackRest delta
restore will be attempted.  This allows users to leverage existing
PGDATA directories when restoring from an existing pgBackRest backup
during cluster initialization, specifically by leveraging
pgBackRest's delta restore capability to cleanly and efficiently
populate the PGDATA directory.

It should be noted that as of this commit, an existing PGATA directory
will only be loaded if the PGHA_BOOTSTRAP_METHOD is set to
'existing_init' (as will automatically occur if PGHA_BOOTSTRAP_METHOD
is not explicitly set and an existing PostgreSQL database is detected).
Therefore, if a user requests to restore from an existing or former
PostgreSQL cluster when bootstrapping a new cluster, but also has
an existing database within the PGDATA directory, the restore will now
now always take precedence (specifically in the form of a pgBackRest
delta restore), overwriting any contents of the PGDATA directory as
needed to complete the delta restore.  This is a change from previous
behavior, in which loading an existing PostgreSQL database (if
detected within the PGDATA directory) would always take precedence over
performing a restore, even if a restore was explicitly requested.

Issue: [ch9814]