Prevent Directory Not Found Errors
When the 'crunchy-postgres-ha' container checks to see if the PGDATA
directory already has data within it during container start up, a
check is now also done to first verify that the PGDATA directory
exists.  This prevents "No such file or directory" errors from
appearing in the container logs when bootstrapping a brand new cluster
using a new PGDATA volume, which might not yet have the proper PGDATA
directory created within it.