Containers Compaction: pgBouncer / pgBadger
In an effort to remove the crunchy-pg-base image, the crunchy-pgbouncer
and crunchy-pgbadger iamges are now built based on the crunchy-base
image. Becaise neither pgBouncer nor pgBadger depend on having
PostgreSQL installed, using the crunchy-pg-base image was making the
images larger than they needed to be. This also removes two more
dependencies on the crunchy-pg-base image.

The scripts for crunchy-pgbouncer and crunchy-pgbadger are updated with
the new `/opt/crunchy` (previously `/opt/cpm`) path in the binary and
config files. The Dockerfiles for these images now include steps to
setup the required repository files and packager settings based on the
base OS, which were previously done in the crunchy-pg-base image.

This also corrects situations where the pgBouncer Dockerfile was
installing an EPEL version of pgBouncer.

Issue: [ch9428]