Update pgpool base image
In an effort to remove the crunchy-pg-base image, the crunchy-pgpool image is
now built based on the crunchy-base. Because pgpool does not need for postgres
to be installed, having crunchy-pg-base as its base image was making the image
larger than it needed to be.

The scripts for crunchy-pgpool have been updated with the new `/opt/crunchy`
(previously `/opt/cpm`) path in the binary and config files. The Dockerfiles
for these images now include steps to setup the required repository files and
packager settings based on the base OS, which were previously done in the
crunchy-pg-base image.

Issue: [ch9428]