Update upgrade container to include pgnodemx
The pgnodemx package has been added to the crunchy-postgres image and is now
included in the default postgresql.conf as an included shared object.
pg_upgrade needs access to any custom shared object files that are installed in
the "old" database. The upgrade image now includes the pgnodemx package for all
versions of postgres that are supported by crunchy repos.