Container Compaction: Postgres Appdev
This change updates the crunchy-postgres-appdev container to be based off of
the crunchy-base image instead of crunchy-pg-base. This is one step closer to
removing the crunchy-pg-base image entirely.

As part of previous compaction changes the binary and config paths were updated
to use `/opt/crunchy` instead of `/opt/cpm`. In order for the appdev container
to build and run with the updated crunchy-postgres scripts the image needed to
install the correct version of the pgrouting and pgnodemx packages. The
postgresql.conf file expects that the image includes the pgnodemx shared
library.