Containers Compaction: Remove crunchy-pg-base
Each container that used the crunchy-pg-base image has been updated to be based
on the crunchy-base image. The steps included in the crunchy-pg-base image have
been moved into either the crunchy-base-image or into the final image.

This commit includes cleanup changes in the Makefile. References to the pg-base
image have been removed.