Ensure proper directory ownership for pgBackRest config
Since 4.6.0 introduced the pgBackRest repository layering on
top of the pgBackRest image, one of the ownership sets was not
being set properly and could potentially cause issues in
more advanced UID modes. In reality, it was never being set
properly, but was fixed in a later line of the code. As we
are using different base images, the fix no longer applied.

Issue: CrunchyData/postgres-operator#2284