Ensure "archive_mode" is enabled for restores in place
When restoring a PostgreSQL cluster that will be utilizing the same
pgBackRest repository (e.g. a "restore in place"), pgBackRest needs to
have "archive_mode" set to "on" in order to properly push history and
other files.

This is in contrast to "archive_mode" being set to "off" when restoring
to a cluster that will use a new pgBackRest repository. This had been
the only setting in the restore process, which did not account for the
"in place" scenario.

Issue: [ch10506]