Add the nss_wrapper for OCP "restricted" SCC support
Adds and enables the nss_wrapper for all containers using UID
entrypoint scripts, specifically as needed for compatibility with the
'restricted' SCC in OpenShift (and therefore the random UID's that are
assigned to containers by OpenShift).  This is done in accordance with
RedHat guidance (see https://access.redhat.com/articles/4859371),
specifically replacing the current solution which directly modifies the
/etc/passwd and /etc/group files.

With this change, the nss_wrapper will now always be enabled when an
entrypoint script (e.g. uid_postgres.sh, uid_pgbackrest.sh, etc.) is
called.  Additionally, the /etc/passwd and /etc/group files are no
longer writeable, and entrypoint script naming has been maintained to
ensure a seamless and transparent transition to the nss_wrapper
solution (i.e. consumers of the containers are not required to make any
changes to maintain current behavior and functionality).

Issue: [ch9887]