Refactors pgo-backrest Util
Refactors the pgo-backrest util as used to exec into containers and run
pgBackRest commands.  This includes better organizing the logic for
obtaining configuration settings and forming the pgBackRest command.
Additionally, now includes the ability to verify the hashes of the
pgBackRest 'conf.d' directory before running the command, while now
also allowing for the discovery of the Pod to be exec'd into via
labels.