Clean and/or Re-create Files on PGHA Init
Various files created and/or utilized when initializing the
"crunchy-postgres-ha" are now properly cleaned and/or recreated when
the various initialize scripts are run to initialize the container
(and therefore Patroni and PostgreSQL).  This ensures various config
files are created with the proper contents if they already exist (e.g.
due to a restart of the container), while also ensuring the log files
do not continuously grow.

Additionally, the nss_wrapper init script will no longer display an
error if the nss_wrapper directory already exists.