Remove the /var/spool/pgbackrest Directory
The default pgBackRest spool directory ("/var/spool/pgbackrest") is now
removed when building the "crunchy-postgres" and "crunchy-pgbackrest"
containers.

This is done to prevent pgBackRest from looking in this directory when
performing a restore, since it was not always have permissions to
access this directory in all environments (e.g. OpenShift).

Issue: [sc-12726]