Fix PG Installs in UBI 8 Upgrade Container
Provides a solution for installing all required PG packages in the
UBI 8 crunchy-upgrade container, specifically to address a limitation
in microdnf in which it is not possible to use regular expressions in
package names when installing a package using "microdnf install".

This solution maintains existing build behavior in which the versions
of PG installed in the upgrade container are determined by the .repo
files present (and therefore included in the container) when building.

[sc-12978]